2748 : 피보나치 수 2

+++ Baekjoon/DynamicProgramming/2748 피보나치 수 2/2748.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main() {
     int n;
     cin >> n;
-    vector<int> vec(n+1, 0);
+    vector<long long> vec(n+1, 0);
 
     vec[0] = 0; vec[1] = 1;
